_RPA_AUTOMATION/gema0/rpa_workers/requirements.txt
pyautogui>=0.9.54
pygetwindow>=0.0.9
pywinauto>=0.6.8
pyperclip>=1.9.0
keyboard>=0.13.5
psutil>=5.9.8
Pillow>=10.3.0
rapidfuzz>=3.9.0
pydantic>=2.7.0
loguru>=0.7.2
